Make the analysis tool engine agnostic
By breaking the old depthmapXNet to a separate class (DepthmapNetEngine) which handles the depthmapXNet connectivity with sockets and a base class (DepthmapEngine) which handles the data parsing and conversion. Also added a generic AnalysisEngine class which currently only contains a common exception and a results storage class but which could be used to introduce non-depthmap engines
